ref(tools): migrate to colocated tests
