+++ packages/core/src/prompts/multi-select.test.ts
@@ -1,6 +1,6 @@
 import { cursor } from 'sisteransi';
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
-import { default as MultiSelectPrompt } from '../../src/prompts/multi-select.js';
+import { default as MultiSelectPrompt } from './multi-select.js';
 import { MockReadable } from '../mock-readable.js';
 import { MockWritable } from '../mock-writable.js';
