+++ packages/core/src/prompts/password.test.ts
@@ -1,7 +1,7 @@
 import { styleText } from 'node:util';
 import { cursor } from 'sisteransi';
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
-import { default as PasswordPrompt } from '../../src/prompts/password.js';
+import { default as PasswordPrompt } from './password.js';
 import { MockReadable } from '../mock-readable.js';
 import { MockWritable } from '../mock-writable.js';
