+++ packages/core/src/prompts/prompt.test.ts
@@ -1,7 +1,7 @@
 import { cursor } from 'sisteransi';
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
-import { default as Prompt } from '../../src/prompts/prompt.js';
-import { isCancel } from '../../src/utils/index.js';
+import { default as Prompt } from './prompt.js';
+import { isCancel } from '../utils/index.js';
 import { MockReadable } from '../mock-readable.js';
 import { MockWritable } from '../mock-writable.js';
