+++ packages/core/src/prompts/select.test.ts
@@ -1,6 +1,6 @@
 import { cursor } from 'sisteransi';
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
-import { default as SelectPrompt } from '../../src/prompts/select.js';
+import { default as SelectPrompt } from './select.js';
 import { MockReadable } from '../mock-readable.js';
 import { MockWritable } from '../mock-writable.js';
