+++ packages/core/src/prompts/text.test.ts
@@ -1,7 +1,7 @@
 import { styleText } from 'node:util';
 import { cursor } from 'sisteransi';
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
-import { default as TextPrompt } from '../../src/prompts/text.js';
+import { default as TextPrompt } from './text.js';
 import { MockReadable } from '../mock-readable.js';
 import { MockWritable } from '../mock-writable.js';
