+++ packages/core/src/test-utils.ts
@@ -1,21 +1,5 @@
 import { Readable, Writable } from 'node:stream';
 
-export class MockWritable extends Writable {
-	public buffer: string[] = [];
-	public isTTY = false;
-	public columns = 80;
-	public rows = 20;
-
-	_write(
-		chunk: any,
-		_encoding: BufferEncoding,
-		callback: (error?: Error | null | undefined) => void
-	): void {
-		this.buffer.push(chunk.toString());
-		callback();
-	}
-}
-
 export class MockReadable extends Readable {
 	protected _buffer: unknown[] | null = [];
 
@@ -38,5 +22,21 @@ export class MockReadable extends Readable {
 
 	close(): void {
 		this._buffer = null;
+	}
+}
+
+export class MockWritable extends Writable {
+	public buffer: string[] = [];
+	public isTTY = false;
+	public columns = 80;
+	public rows = 20;
+
+	_write(
+		chunk: any,
+		_encoding: BufferEncoding,
+		callback: (error?: Error | null | undefined) => void
+	): void {
+		this.buffer.push(chunk.toString());
+		callback();
 	}
 }
