+++ packages/prompts/package.json
@@ -48,7 +48,7 @@
   ],
   "packageManager": "pnpm@9.14.2",
   "scripts": {
-    "build": "bsh build",
+    "build": "bsh build \"src/**/*.ts\" \"!src/**/*.test.ts\"",
     "prepack": "pnpm build",
     "test": "bsh test",
     "lint": "bsh lint"
