+++ packages/prompts/src/box.test.ts
@@ -1,7 +1,7 @@
 import { styleText } from 'node:util';
 import { updateSettings } from '@clack/core';
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test, vi } from 'vitest';
-import * as prompts from '../src/index.js';
+import * as prompts from './index.js';
 import { MockReadable, MockWritable } from './test-utils.js';
 
 describe.each(['true', 'false'])('box (isCI = %s)', (isCI) => {
