+++ packages/prompts/src/date.test.ts
@@ -1,6 +1,6 @@
 import { updateSettings } from '@clack/core';
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test, vi } from 'vitest';
-import * as prompts from '../src/index.js';
+import * as prompts from './index.js';
 import { MockReadable, MockWritable } from './test-utils.js';
 
 const d = (iso: string) => {
