+++ packages/prompts/src/note.test.ts
@@ -1,6 +1,6 @@
 import { styleText } from 'node:util';
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test, vi } from 'vitest';
-import * as prompts from '../src/index.js';
+import * as prompts from './index.js';
 import { MockReadable, MockWritable } from './test-utils.js';
 
 describe.each(['true', 'false'])('note (isCI = %s)', (isCI) => {
