+++ packages/prompts/src/progress-bar.test.ts
@@ -1,11 +1,11 @@
+import process from 'node:process';
 import { EventEmitter } from 'node:stream';
-import { styleText } from 'node:util';
-import { getColumns, updateSettings } from '@clack/core';
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test, vi } from 'vitest';
-import * as prompts from '../src/index.js';
+import type { ProgressOptions } from './index.js';
+import * as prompts from './index.js';
 import { MockWritable } from './test-utils.js';
 
-describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
+describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 	let originalCI: string | undefined;
 	let output: MockWritable;
 
@@ -24,22 +24,22 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 	});
 
 	afterEach(() => {
-		vi.useRealTimers();
 		vi.restoreAllMocks();
-		updateSettings({ withGuide: true });
+		vi.useRealTimers();
 	});
 
-	test('returns spinner API', () => {
-		const api = prompts.spinner({ output });
+	test('returns progress API', () => {
+		const api = prompts.progress({ output });
 
 		expect(api.stop).toBeTypeOf('function');
 		expect(api.start).toBeTypeOf('function');
 		expect(api.message).toBeTypeOf('function');
+		expect(api.advance).toBeTypeOf('function');
 	});
 
 	describe('start', () => {
 		test('renders frames at interval', () => {
-			const result = prompts.spinner({ output });
+			const result = prompts.progress({ output });
 
 			result.start();
 
@@ -48,64 +48,33 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 				vi.advanceTimersByTime(80);
 			}
 
-			result.stop();
-
 			expect(output.buffer).toMatchSnapshot();
 		});
 
 		test('renders message', () => {
-			const result = prompts.spinner({ output });
+			const result = prompts.progress({ output });
 
 			result.start('foo');
 
 			vi.advanceTimersByTime(80);
 
-			result.stop();
-
 			expect(output.buffer).toMatchSnapshot();
 		});
 
 		test('renders timer when indicator is "timer"', () => {
-			const result = prompts.spinner({ output, indicator: 'timer' });
+			const result = prompts.progress({ output, indicator: 'timer' });
 
 			result.start();
 
 			vi.advanceTimersByTime(80);
-
-			result.stop();
-
-			expect(output.buffer).toMatchSnapshot();
-		});
-
-		test('handles wrapping', () => {
-			const columns = getColumns(output);
-			const result = prompts.spinner({ output });
-
-			result.start('x'.repeat(columns + 10));
-
-			vi.advanceTimersByTime(80);
-
-			result.stop('stopped');
-
-			expect(output.buffer).toMatchSnapshot();
-		});
-
-		test('handles multi-line messages', () => {
-			const result = prompts.spinner({ output });
-
-			result.start('foo\nbar\nbaz');
-
-			vi.advanceTimersByTime(80);
-
-			result.stop();
 
 			expect(output.buffer).toMatchSnapshot();
 		});
 	});
 
 	describe('stop', () => {
-		test('renders submit symbol and stops spinner', () => {
-			const result = prompts.spinner({ output });
+		test('renders submit symbol and stops progress', () => {
+			const result = prompts.progress({ output });
 
 			result.start();
 
@@ -119,7 +88,7 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 		});
 
 		test('renders cancel symbol when calling cancel()', () => {
-			const result = prompts.spinner({ output });
+			const result = prompts.progress({ output });
 
 			result.start();
 
@@ -131,7 +100,7 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 		});
 
 		test('renders error symbol when calling error()', () => {
-			const result = prompts.spinner({ output });
+			const result = prompts.progress({ output });
 
 			result.start();
 
@@ -143,7 +112,7 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 		});
 
 		test('renders message', () => {
-			const result = prompts.spinner({ output });
+			const result = prompts.progress({ output });
 
 			result.start();
 
@@ -155,7 +124,7 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 		});
 
 		test('renders message without removing dots', () => {
-			const result = prompts.spinner({ output });
+			const result = prompts.progress({ output });
 
 			result.start();
 
@@ -167,39 +136,33 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 		});
 
 		test('renders message when cancelling', () => {
-			const result = prompts.spinner({ output });
+			const result = prompts.progress({ output });
 
 			result.start();
 
 			vi.advanceTimersByTime(80);
 
-			result.cancel('too dizzy — spinning cancelled');
+			result.cancel('cancelled :-(');
 
 			expect(output.buffer).toMatchSnapshot();
 		});
 
 		test('renders message when erroring', () => {
-			const result = prompts.spinner({ output });
+			const result = prompts.progress({ output });
 
 			result.start();
 
 			vi.advanceTimersByTime(80);
 
-			result.error('error: spun too fast!');
+			result.error('FATAL ERROR!');
 
 			expect(output.buffer).toMatchSnapshot();
-		});
-
-		test('does not throw if called before start', () => {
-			const result = prompts.spinner({ output });
-
-			expect(() => result.stop()).not.toThrow();
 		});
 	});
 
 	describe('message', () => {
 		test('sets message for next frame', () => {
-			const result = prompts.spinner({ output });
+			const result = prompts.progress({ output });
 
 			result.start();
 
@@ -208,68 +171,6 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 			result.message('foo');
 
 			vi.advanceTimersByTime(80);
-
-			result.stop();
-
-			expect(output.buffer).toMatchSnapshot();
-		});
-	});
-
-	describe('indicator customization', () => {
-		test('custom frames', () => {
-			const result = prompts.spinner({ output, frames: ['🐴', '🦋', '🐙', '🐶'] });
-
-			result.start();
-
-			// there are 4 frames
-			for (let i = 0; i < 4; i++) {
-				vi.advanceTimersByTime(80);
-			}
-
-			result.stop();
-
-			expect(output.buffer).toMatchSnapshot();
-		});
-
-		test('custom frames with lots of frame have consistent ellipsis display', () => {
-			const result = prompts.spinner({ output, frames: Object.keys(Array(10).fill(0)) });
-
-			result.start();
-
-			for (let i = 0; i < 64; i++) {
-				vi.advanceTimersByTime(80);
-			}
-
-			result.stop();
-
-			expect(output.buffer).toMatchSnapshot();
-		});
-
-		test('custom delay', () => {
-			const result = prompts.spinner({ output, delay: 200 });
-
-			result.start();
-
-			// there are 4 frames
-			for (let i = 0; i < 4; i++) {
-				vi.advanceTimersByTime(200);
-			}
-
-			result.stop();
-
-			expect(output.buffer).toMatchSnapshot();
-		});
-
-		test('custom frame style', () => {
-			const result = prompts.spinner({ output, styleFrame: (text) => styleText('red', text) });
-
-			result.start();
-
-			for (let i = 0; i < 4; i++) {
-				vi.advanceTimersByTime(80);
-			}
-
-			result.stop();
 
 			expect(output.buffer).toMatchSnapshot();
 		});
@@ -297,7 +198,7 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 		});
 
 		test('uses default cancel message', () => {
-			const result = prompts.spinner({ output });
+			const result = prompts.progress({ output });
 			result.start('Test operation');
 
 			processEmitter.emit('SIGINT');
@@ -306,7 +207,7 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 		});
 
 		test('uses custom cancel message when provided directly', () => {
-			const result = prompts.spinner({
+			const result = prompts.progress({
 				output,
 				cancelMessage: 'Custom cancel message',
 			});
@@ -318,7 +219,7 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 		});
 
 		test('uses custom error message when provided directly', () => {
-			const result = prompts.spinner({
+			const result = prompts.progress({
 				output,
 				errorMessage: 'Custom error message',
 			});
@@ -336,7 +237,7 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 				// Set custom message
 				prompts.settings.messages.cancel = 'Global cancel message';
 
-				const result = prompts.spinner({ output });
+				const result = prompts.progress({ output });
 				result.start('Test operation');
 
 				processEmitter.emit('SIGINT');
@@ -351,12 +252,11 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 		test('uses global custom error message from settings', () => {
 			// Store original message
 			const originalErrorMessage = prompts.settings.messages.error;
-
 			try {
 				// Set custom message
 				prompts.settings.messages.error = 'Global error message';
 
-				const result = prompts.spinner({ output });
+				const result = prompts.progress({ output });
 				result.start('Test operation');
 
 				processEmitter.emit('exit', 2);
@@ -376,9 +276,9 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 				// Set custom global messages
 				prompts.settings.messages.error = 'Global error message';
 
-				const result = prompts.spinner({
+				const result = prompts.progress({
 					output,
-					errorMessage: 'Spinner error message',
+					errorMessage: 'Progress error message',
 				});
 				result.start('Test operation');
 
@@ -398,9 +298,9 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 				// Set custom global messages
 				prompts.settings.messages.cancel = 'Global cancel message';
 
-				const result = prompts.spinner({
+				const result = prompts.progress({
 					output,
-					cancelMessage: 'Spinner cancel message',
+					cancelMessage: 'Progress cancel message',
 				});
 				result.start('Test operation');
 
@@ -413,57 +313,19 @@ describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 		});
 	});
 
-	test('can be aborted by a signal', async () => {
-		const controller = new AbortController();
-		const result = prompts.spinner({
-			output,
-			signal: controller.signal,
-		});
+	describe('style', () => {
+		test.each(['block', 'heavy', 'light'] satisfies Array<ProgressOptions['style']>)(
+			'renders %s progressbar',
+			(style) => {
+				const result = prompts.progress({ output, style, max: 2, size: 10 });
+				result.start();
+				vi.advanceTimersByTime(160);
+				result.advance();
+				vi.advanceTimersByTime(160);
+				result.stop();
 
-		result.start('Testing');
-
-		controller.abort();
-
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('withGuide: false removes guide', () => {
-		const result = prompts.spinner({ output, withGuide: false });
-
-		result.start('foo');
-
-		vi.advanceTimersByTime(80);
-
-		result.stop();
-
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('global withGuide: false removes guide', () => {
-		updateSettings({ withGuide: false });
-
-		const result = prompts.spinner({ output });
-
-		result.start('foo');
-
-		vi.advanceTimersByTime(80);
-
-		result.stop();
-
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	describe('clear', () => {
-		test('stops and clears the spinner from the output', () => {
-			const result = prompts.spinner({ output });
-
-			result.start('Loading');
-
-			vi.advanceTimersByTime(80);
-
-			result.clear();
-
-			expect(output.buffer).toMatchSnapshot();
-		});
+				expect(output.buffer).toMatchSnapshot();
+			}
+		);
 	});
 });
