+++ packages/prompts/src/spinner.test.ts
@@ -1,11 +1,11 @@
-import process from 'node:process';
 import { EventEmitter } from 'node:stream';
+import { styleText } from 'node:util';
+import { getColumns, updateSettings } from '@clack/core';
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test, vi } from 'vitest';
-import type { ProgressOptions } from '../src/index.js';
-import * as prompts from '../src/index.js';
+import * as prompts from './index.js';
 import { MockWritable } from './test-utils.js';
 
-describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
+describe.each(['true', 'false'])('spinner (isCI = %s)', (isCI) => {
 	let originalCI: string | undefined;
 	let output: MockWritable;
 
@@ -24,22 +24,22 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 	});
 
 	afterEach(() => {
-		vi.restoreAllMocks();
 		vi.useRealTimers();
+		vi.restoreAllMocks();
+		updateSettings({ withGuide: true });
 	});
 
-	test('returns progress API', () => {
-		const api = prompts.progress({ output });
+	test('returns spinner API', () => {
+		const api = prompts.spinner({ output });
 
 		expect(api.stop).toBeTypeOf('function');
 		expect(api.start).toBeTypeOf('function');
 		expect(api.message).toBeTypeOf('function');
-		expect(api.advance).toBeTypeOf('function');
 	});
 
 	describe('start', () => {
 		test('renders frames at interval', () => {
-			const result = prompts.progress({ output });
+			const result = prompts.spinner({ output });
 
 			result.start();
 
@@ -48,33 +48,64 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 				vi.advanceTimersByTime(80);
 			}
 
+			result.stop();
+
 			expect(output.buffer).toMatchSnapshot();
 		});
 
 		test('renders message', () => {
-			const result = prompts.progress({ output });
+			const result = prompts.spinner({ output });
 
 			result.start('foo');
 
 			vi.advanceTimersByTime(80);
 
+			result.stop();
+
 			expect(output.buffer).toMatchSnapshot();
 		});
 
 		test('renders timer when indicator is "timer"', () => {
-			const result = prompts.progress({ output, indicator: 'timer' });
+			const result = prompts.spinner({ output, indicator: 'timer' });
 
 			result.start();
 
 			vi.advanceTimersByTime(80);
+
+			result.stop();
+
+			expect(output.buffer).toMatchSnapshot();
+		});
+
+		test('handles wrapping', () => {
+			const columns = getColumns(output);
+			const result = prompts.spinner({ output });
+
+			result.start('x'.repeat(columns + 10));
+
+			vi.advanceTimersByTime(80);
+
+			result.stop('stopped');
+
+			expect(output.buffer).toMatchSnapshot();
+		});
+
+		test('handles multi-line messages', () => {
+			const result = prompts.spinner({ output });
+
+			result.start('foo\nbar\nbaz');
+
+			vi.advanceTimersByTime(80);
+
+			result.stop();
 
 			expect(output.buffer).toMatchSnapshot();
 		});
 	});
 
 	describe('stop', () => {
-		test('renders submit symbol and stops progress', () => {
-			const result = prompts.progress({ output });
+		test('renders submit symbol and stops spinner', () => {
+			const result = prompts.spinner({ output });
 
 			result.start();
 
@@ -88,7 +119,7 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 		});
 
 		test('renders cancel symbol when calling cancel()', () => {
-			const result = prompts.progress({ output });
+			const result = prompts.spinner({ output });
 
 			result.start();
 
@@ -100,7 +131,7 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 		});
 
 		test('renders error symbol when calling error()', () => {
-			const result = prompts.progress({ output });
+			const result = prompts.spinner({ output });
 
 			result.start();
 
@@ -112,7 +143,7 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 		});
 
 		test('renders message', () => {
-			const result = prompts.progress({ output });
+			const result = prompts.spinner({ output });
 
 			result.start();
 
@@ -124,7 +155,7 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 		});
 
 		test('renders message without removing dots', () => {
-			const result = prompts.progress({ output });
+			const result = prompts.spinner({ output });
 
 			result.start();
 
@@ -136,33 +167,39 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 		});
 
 		test('renders message when cancelling', () => {
-			const result = prompts.progress({ output });
+			const result = prompts.spinner({ output });
 
 			result.start();
 
 			vi.advanceTimersByTime(80);
 
-			result.cancel('cancelled :-(');
+			result.cancel('too dizzy — spinning cancelled');
 
 			expect(output.buffer).toMatchSnapshot();
 		});
 
 		test('renders message when erroring', () => {
-			const result = prompts.progress({ output });
+			const result = prompts.spinner({ output });
 
 			result.start();
 
 			vi.advanceTimersByTime(80);
 
-			result.error('FATAL ERROR!');
+			result.error('error: spun too fast!');
 
 			expect(output.buffer).toMatchSnapshot();
+		});
+
+		test('does not throw if called before start', () => {
+			const result = prompts.spinner({ output });
+
+			expect(() => result.stop()).not.toThrow();
 		});
 	});
 
 	describe('message', () => {
 		test('sets message for next frame', () => {
-			const result = prompts.progress({ output });
+			const result = prompts.spinner({ output });
 
 			result.start();
 
@@ -171,6 +208,68 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 			result.message('foo');
 
 			vi.advanceTimersByTime(80);
+
+			result.stop();
+
+			expect(output.buffer).toMatchSnapshot();
+		});
+	});
+
+	describe('indicator customization', () => {
+		test('custom frames', () => {
+			const result = prompts.spinner({ output, frames: ['🐴', '🦋', '🐙', '🐶'] });
+
+			result.start();
+
+			// there are 4 frames
+			for (let i = 0; i < 4; i++) {
+				vi.advanceTimersByTime(80);
+			}
+
+			result.stop();
+
+			expect(output.buffer).toMatchSnapshot();
+		});
+
+		test('custom frames with lots of frame have consistent ellipsis display', () => {
+			const result = prompts.spinner({ output, frames: Object.keys(Array(10).fill(0)) });
+
+			result.start();
+
+			for (let i = 0; i < 64; i++) {
+				vi.advanceTimersByTime(80);
+			}
+
+			result.stop();
+
+			expect(output.buffer).toMatchSnapshot();
+		});
+
+		test('custom delay', () => {
+			const result = prompts.spinner({ output, delay: 200 });
+
+			result.start();
+
+			// there are 4 frames
+			for (let i = 0; i < 4; i++) {
+				vi.advanceTimersByTime(200);
+			}
+
+			result.stop();
+
+			expect(output.buffer).toMatchSnapshot();
+		});
+
+		test('custom frame style', () => {
+			const result = prompts.spinner({ output, styleFrame: (text) => styleText('red', text) });
+
+			result.start();
+
+			for (let i = 0; i < 4; i++) {
+				vi.advanceTimersByTime(80);
+			}
+
+			result.stop();
 
 			expect(output.buffer).toMatchSnapshot();
 		});
@@ -198,7 +297,7 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 		});
 
 		test('uses default cancel message', () => {
-			const result = prompts.progress({ output });
+			const result = prompts.spinner({ output });
 			result.start('Test operation');
 
 			processEmitter.emit('SIGINT');
@@ -207,7 +306,7 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 		});
 
 		test('uses custom cancel message when provided directly', () => {
-			const result = prompts.progress({
+			const result = prompts.spinner({
 				output,
 				cancelMessage: 'Custom cancel message',
 			});
@@ -219,7 +318,7 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 		});
 
 		test('uses custom error message when provided directly', () => {
-			const result = prompts.progress({
+			const result = prompts.spinner({
 				output,
 				errorMessage: 'Custom error message',
 			});
@@ -237,7 +336,7 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 				// Set custom message
 				prompts.settings.messages.cancel = 'Global cancel message';
 
-				const result = prompts.progress({ output });
+				const result = prompts.spinner({ output });
 				result.start('Test operation');
 
 				processEmitter.emit('SIGINT');
@@ -252,11 +351,12 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 		test('uses global custom error message from settings', () => {
 			// Store original message
 			const originalErrorMessage = prompts.settings.messages.error;
+
 			try {
 				// Set custom message
 				prompts.settings.messages.error = 'Global error message';
 
-				const result = prompts.progress({ output });
+				const result = prompts.spinner({ output });
 				result.start('Test operation');
 
 				processEmitter.emit('exit', 2);
@@ -276,9 +376,9 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 				// Set custom global messages
 				prompts.settings.messages.error = 'Global error message';
 
-				const result = prompts.progress({
+				const result = prompts.spinner({
 					output,
-					errorMessage: 'Progress error message',
+					errorMessage: 'Spinner error message',
 				});
 				result.start('Test operation');
 
@@ -298,9 +398,9 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 				// Set custom global messages
 				prompts.settings.messages.cancel = 'Global cancel message';
 
-				const result = prompts.progress({
+				const result = prompts.spinner({
 					output,
-					cancelMessage: 'Progress cancel message',
+					cancelMessage: 'Spinner cancel message',
 				});
 				result.start('Test operation');
 
@@ -313,19 +413,57 @@ describe.each(['true', 'false'])('prompts - progress (isCI = %s)', (isCI) => {
 		});
 	});
 
-	describe('style', () => {
-		test.each(['block', 'heavy', 'light'] satisfies Array<ProgressOptions['style']>)(
-			'renders %s progressbar',
-			(style) => {
-				const result = prompts.progress({ output, style, max: 2, size: 10 });
-				result.start();
-				vi.advanceTimersByTime(160);
-				result.advance();
-				vi.advanceTimersByTime(160);
-				result.stop();
+	test('can be aborted by a signal', async () => {
+		const controller = new AbortController();
+		const result = prompts.spinner({
+			output,
+			signal: controller.signal,
+		});
 
-				expect(output.buffer).toMatchSnapshot();
-			}
-		);
+		result.start('Testing');
+
+		controller.abort();
+
+		expect(output.buffer).toMatchSnapshot();
+	});
+
+	test('withGuide: false removes guide', () => {
+		const result = prompts.spinner({ output, withGuide: false });
+
+		result.start('foo');
+
+		vi.advanceTimersByTime(80);
+
+		result.stop();
+
+		expect(output.buffer).toMatchSnapshot();
+	});
+
+	test('global withGuide: false removes guide', () => {
+		updateSettings({ withGuide: false });
+
+		const result = prompts.spinner({ output });
+
+		result.start('foo');
+
+		vi.advanceTimersByTime(80);
+
+		result.stop();
+
+		expect(output.buffer).toMatchSnapshot();
+	});
+
+	describe('clear', () => {
+		test('stops and clears the spinner from the output', () => {
+			const result = prompts.spinner({ output });
+
+			result.start('Loading');
+
+			vi.advanceTimersByTime(80);
+
+			result.clear();
+
+			expect(output.buffer).toMatchSnapshot();
+		});
 	});
 });
