+++ packages/prompts/src/text.test.ts
@@ -1,9 +1,9 @@
 import { updateSettings } from '@clack/core';
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test, vi } from 'vitest';
-import * as prompts from '../src/index.js';
+import * as prompts from './index.js';
 import { MockReadable, MockWritable } from './test-utils.js';
 
-describe.each(['true', 'false'])('confirm (isCI = %s)', (isCI) => {
+describe.each(['true', 'false'])('text (isCI = %s)', (isCI) => {
 	let originalCI: string | undefined;
 	let output: MockWritable;
 	let input: MockReadable;
@@ -27,8 +27,8 @@ describe.each(['true', 'false'])('confirm (isCI = %s)', (isCI) => {
 		updateSettings({ withGuide: true });
 	});
 
-	test('renders message with choices', async () => {
-		const result = prompts.confirm({
+	test('renders message', async () => {
+		const result = prompts.text({
 			message: 'foo',
 			input,
 			output,
@@ -36,16 +36,15 @@ describe.each(['true', 'false'])('confirm (isCI = %s)', (isCI) => {
 
 		input.emit('keypress', '', { name: 'return' });
 
-		const value = await result;
+		await result;
 
-		expect(value).toBe(true);
 		expect(output.buffer).toMatchSnapshot();
 	});
 
-	test('renders custom active choice', async () => {
-		const result = prompts.confirm({
+	test('renders placeholder if set', async () => {
+		const result = prompts.text({
 			message: 'foo',
-			active: 'bleep',
+			placeholder: 'bar',
 			input,
 			output,
 		});
@@ -54,77 +53,12 @@ describe.each(['true', 'false'])('confirm (isCI = %s)', (isCI) => {
 
 		const value = await result;
 
-		expect(value).toBe(true);
 		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('renders custom inactive choice', async () => {
-		const result = prompts.confirm({
-			message: 'foo',
-			inactive: 'bleep',
-			input,
-			output,
-		});
-
-		input.emit('keypress', '', { name: 'return' });
-
-		const value = await result;
-
-		expect(value).toBe(true);
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('renders options in vertical alignment', async () => {
-		const result = prompts.confirm({
-			message: 'foo',
-			vertical: true,
-			input,
-			output,
-		});
-
-		input.emit('keypress', '', { name: 'return' });
-
-		const value = await result;
-
-		expect(value).toBe(true);
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('right arrow moves to next choice', async () => {
-		const result = prompts.confirm({
-			message: 'foo',
-			input,
-			output,
-		});
-
-		input.emit('keypress', 'right', { name: 'right' });
-		input.emit('keypress', '', { name: 'return' });
-
-		const value = await result;
-
-		expect(value).toBe(false);
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('left arrow moves to previous choice', async () => {
-		const result = prompts.confirm({
-			message: 'foo',
-			input,
-			output,
-		});
-
-		input.emit('keypress', 'right', { name: 'right' });
-		input.emit('keypress', 'left', { name: 'left' });
-		input.emit('keypress', '', { name: 'return' });
-
-		const value = await result;
-
-		expect(value).toBe(true);
-		expect(output.buffer).toMatchSnapshot();
+		expect(value).toBe('');
 	});
 
 	test('can cancel', async () => {
-		const result = prompts.confirm({
+		const result = prompts.text({
 			message: 'foo',
 			input,
 			output,
@@ -138,10 +72,44 @@ describe.each(['true', 'false'])('confirm (isCI = %s)', (isCI) => {
 		expect(output.buffer).toMatchSnapshot();
 	});
 
-	test('can set initialValue', async () => {
-		const result = prompts.confirm({
+	test('renders cancelled value if one set', async () => {
+		const result = prompts.text({
 			message: 'foo',
-			initialValue: false,
+			input,
+			output,
+		});
+
+		input.emit('keypress', 'x', { name: 'x' });
+		input.emit('keypress', 'y', { name: 'y' });
+		input.emit('keypress', '', { name: 'escape' });
+
+		const value = await result;
+
+		expect(prompts.isCancel(value)).toBe(true);
+		expect(output.buffer).toMatchSnapshot();
+	});
+
+	test('renders submitted value', async () => {
+		const result = prompts.text({
+			message: 'foo',
+			input,
+			output,
+		});
+
+		input.emit('keypress', 'x', { name: 'x' });
+		input.emit('keypress', 'y', { name: 'y' });
+		input.emit('keypress', '', { name: 'return' });
+
+		const value = await result;
+
+		expect(value).toBe('xy');
+		expect(output.buffer).toMatchSnapshot();
+	});
+
+	test('defaultValue sets the value but does not render', async () => {
+		const result = prompts.text({
+			message: 'foo',
+			defaultValue: 'bar',
 			input,
 			output,
 		});
@@ -150,14 +118,85 @@ describe.each(['true', 'false'])('confirm (isCI = %s)', (isCI) => {
 
 		const value = await result;
 
-		expect(value).toBe(false);
+		expect(value).toBe('bar');
+		expect(output.buffer).toMatchSnapshot();
+	});
+
+	test('validation errors render and clear', async () => {
+		const result = prompts.text({
+			message: 'foo',
+			validate: (val) => (val !== 'xy' ? 'should be xy' : undefined),
+			input,
+			output,
+		});
+
+		input.emit('keypress', 'x', { name: 'x' });
+		input.emit('keypress', '', { name: 'return' });
+		input.emit('keypress', 'y', { name: 'y' });
+		input.emit('keypress', '', { name: 'return' });
+
+		const value = await result;
+
+		expect(value).toBe('xy');
+		expect(output.buffer).toMatchSnapshot();
+	});
+
+	test('validation errors render and clear (using Error)', async () => {
+		const result = prompts.text({
+			message: 'foo',
+			validate: (val) => (val !== 'xy' ? new Error('should be xy') : undefined),
+			input,
+			output,
+		});
+
+		input.emit('keypress', 'x', { name: 'x' });
+		input.emit('keypress', '', { name: 'return' });
+		input.emit('keypress', 'y', { name: 'y' });
+		input.emit('keypress', '', { name: 'return' });
+
+		const value = await result;
+
+		expect(value).toBe('xy');
+		expect(output.buffer).toMatchSnapshot();
+	});
+
+	test('placeholder is not used as value when pressing enter', async () => {
+		const result = prompts.text({
+			message: 'foo',
+			placeholder: '  (hit Enter to use default)',
+			defaultValue: 'default-value',
+			input,
+			output,
+		});
+
+		input.emit('keypress', '', { name: 'return' });
+
+		const value = await result;
+
+		expect(value).toBe('default-value');
+		expect(output.buffer).toMatchSnapshot();
+	});
+
+	test('empty string when no value and no default', async () => {
+		const result = prompts.text({
+			message: 'foo',
+			placeholder: '  (hit Enter to use default)',
+			input,
+			output,
+		});
+
+		input.emit('keypress', '', { name: 'return' });
+
+		const value = await result;
+
+		expect(value).toBe('');
 		expect(output.buffer).toMatchSnapshot();
 	});
 
 	test('can be aborted by a signal', async () => {
 		const controller = new AbortController();
-		const result = prompts.confirm({
-			message: 'yes?',
+		const result = prompts.text({
+			message: 'foo',
 			input,
 			output,
 			signal: controller.signal,
@@ -170,7 +209,7 @@ describe.each(['true', 'false'])('confirm (isCI = %s)', (isCI) => {
 	});
 
 	test('withGuide: false removes guide', async () => {
-		const result = prompts.confirm({
+		const result = prompts.text({
 			message: 'foo',
 			withGuide: false,
 			input,
@@ -187,7 +226,7 @@ describe.each(['true', 'false'])('confirm (isCI = %s)', (isCI) => {
 	test('global withGuide: false removes guide', async () => {
 		updateSettings({ withGuide: false });
 
-		const result = prompts.confirm({
+		const result = prompts.text({
 			message: 'foo',
 			input,
 			output,
