--- packages/prompts/test/text.test.ts
@@ -1,241 +1,0 @@
-import { updateSettings } from '@clack/core';
-import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test, vi } from 'vitest';
-import * as prompts from '../src/index.js';
-import { MockReadable, MockWritable } from './test-utils.js';
-
-describe.each(['true', 'false'])('text (isCI = %s)', (isCI) => {
-	let originalCI: string | undefined;
-	let output: MockWritable;
-	let input: MockReadable;
-
-	beforeAll(() => {
-		originalCI = process.env.CI;
-		process.env.CI = isCI;
-	});
-
-	afterAll(() => {
-		process.env.CI = originalCI;
-	});
-
-	beforeEach(() => {
-		output = new MockWritable();
-		input = new MockReadable();
-	});
-
-	afterEach(() => {
-		vi.restoreAllMocks();
-		updateSettings({ withGuide: true });
-	});
-
-	test('renders message', async () => {
-		const result = prompts.text({
-			message: 'foo',
-			input,
-			output,
-		});
-
-		input.emit('keypress', '', { name: 'return' });
-
-		await result;
-
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('renders placeholder if set', async () => {
-		const result = prompts.text({
-			message: 'foo',
-			placeholder: 'bar',
-			input,
-			output,
-		});
-
-		input.emit('keypress', '', { name: 'return' });
-
-		const value = await result;
-
-		expect(output.buffer).toMatchSnapshot();
-		expect(value).toBe('');
-	});
-
-	test('can cancel', async () => {
-		const result = prompts.text({
-			message: 'foo',
-			input,
-			output,
-		});
-
-		input.emit('keypress', 'escape', { name: 'escape' });
-
-		const value = await result;
-
-		expect(prompts.isCancel(value)).toBe(true);
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('renders cancelled value if one set', async () => {
-		const result = prompts.text({
-			message: 'foo',
-			input,
-			output,
-		});
-
-		input.emit('keypress', 'x', { name: 'x' });
-		input.emit('keypress', 'y', { name: 'y' });
-		input.emit('keypress', '', { name: 'escape' });
-
-		const value = await result;
-
-		expect(prompts.isCancel(value)).toBe(true);
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('renders submitted value', async () => {
-		const result = prompts.text({
-			message: 'foo',
-			input,
-			output,
-		});
-
-		input.emit('keypress', 'x', { name: 'x' });
-		input.emit('keypress', 'y', { name: 'y' });
-		input.emit('keypress', '', { name: 'return' });
-
-		const value = await result;
-
-		expect(value).toBe('xy');
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('defaultValue sets the value but does not render', async () => {
-		const result = prompts.text({
-			message: 'foo',
-			defaultValue: 'bar',
-			input,
-			output,
-		});
-
-		input.emit('keypress', '', { name: 'return' });
-
-		const value = await result;
-
-		expect(value).toBe('bar');
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('validation errors render and clear', async () => {
-		const result = prompts.text({
-			message: 'foo',
-			validate: (val) => (val !== 'xy' ? 'should be xy' : undefined),
-			input,
-			output,
-		});
-
-		input.emit('keypress', 'x', { name: 'x' });
-		input.emit('keypress', '', { name: 'return' });
-		input.emit('keypress', 'y', { name: 'y' });
-		input.emit('keypress', '', { name: 'return' });
-
-		const value = await result;
-
-		expect(value).toBe('xy');
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('validation errors render and clear (using Error)', async () => {
-		const result = prompts.text({
-			message: 'foo',
-			validate: (val) => (val !== 'xy' ? new Error('should be xy') : undefined),
-			input,
-			output,
-		});
-
-		input.emit('keypress', 'x', { name: 'x' });
-		input.emit('keypress', '', { name: 'return' });
-		input.emit('keypress', 'y', { name: 'y' });
-		input.emit('keypress', '', { name: 'return' });
-
-		const value = await result;
-
-		expect(value).toBe('xy');
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('placeholder is not used as value when pressing enter', async () => {
-		const result = prompts.text({
-			message: 'foo',
-			placeholder: '  (hit Enter to use default)',
-			defaultValue: 'default-value',
-			input,
-			output,
-		});
-
-		input.emit('keypress', '', { name: 'return' });
-
-		const value = await result;
-
-		expect(value).toBe('default-value');
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('empty string when no value and no default', async () => {
-		const result = prompts.text({
-			message: 'foo',
-			placeholder: '  (hit Enter to use default)',
-			input,
-			output,
-		});
-
-		input.emit('keypress', '', { name: 'return' });
-
-		const value = await result;
-
-		expect(value).toBe('');
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('can be aborted by a signal', async () => {
-		const controller = new AbortController();
-		const result = prompts.text({
-			message: 'foo',
-			input,
-			output,
-			signal: controller.signal,
-		});
-
-		controller.abort();
-		const value = await result;
-		expect(prompts.isCancel(value)).toBe(true);
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('withGuide: false removes guide', async () => {
-		const result = prompts.text({
-			message: 'foo',
-			withGuide: false,
-			input,
-			output,
-		});
-
-		input.emit('keypress', '', { name: 'return' });
-
-		await result;
-
-		expect(output.buffer).toMatchSnapshot();
-	});
-
-	test('global withGuide: false removes guide', async () => {
-		updateSettings({ withGuide: false });
-
-		const result = prompts.text({
-			message: 'foo',
-			input,
-			output,
-		});
-
-		input.emit('keypress', '', { name: 'return' });
-
-		await result;
-
-		expect(output.buffer).toMatchSnapshot();
-	});
-});
