+++ packages/prompts/vitest.config.ts
@@ -1,0 +1,11 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+	test: {
+		exclude: ['dist/**', 'node_modules/**'],
+		env: {
+			FORCE_COLOR: '1',
+		},
+		snapshotSerializers: ['vitest-ansi-serializer'],
+	},
+});
